Implement Maya Light Manager UI and Logic classes for enhanced lighting control
